requirements.txt
#!/usr/bin/env python3
"""
Requirements file for the CV-to-Job Matching System.
"""

# Core dependencies
flask==2.3.3
werkzeug==2.3.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# NLP libraries
spacy==3.6.1
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.30.2

# Document processing
python-docx==0.8.11
PyPDF2==3.0.1
pdfminer.six==20221105

# Utilities
python-dotenv==1.0.0
requests==2.31.0
